test fixtures

+++ WordFinder/src/ch/unibe/scg/team3/localDatabase/SavedGamesHandler.java
@@ -36,7 +36,7 @@ public class SavedGamesHandler extends DataHandler {
 
 		SharedPreferences preferences = PreferenceManager
 				.getDefaultSharedPreferences(context);
-		String wordlist = preferences.getString("choose_wordlist", null);
+		int wordlist = Integer.parseInt(preferences.getString("choose_wordlist", null));
 		int timesPlayed = 1;
 		String date = (new Date().toString());
 		if (!gameInDatabase(name)) {
@@ -139,17 +139,16 @@ public class SavedGamesHandler extends DataHandler {
 	}
 
 	public boolean removeGameByName(String name) {
-		SQLiteDatabase db = helper.getReadableDatabase();
-
-		String[] query = { name };
-		try {
+		if (gameInDatabase(name)){
+			SQLiteDatabase db = helper.getReadableDatabase();
+			String[] query = { name };
 			db.execSQL("DELETE FROM Games WHERE Name = ?", query);
-		} catch (SQLException e) {
 			db.close();
-			return false;
+			return true;
 		}
-		db.close();
-		return true;
+		else{
+		return false;
+		}
 	}
 
 }
